[refactor] #133 UserNotification Live, Test 분리
