+++ Projects/CoreKit/Sources/Data/Client/UserNotificationClient/UserNotificationClient.swift
@@ -5,6 +5,7 @@
 //  Created by 김민호 on 7/4/24.
 //
 
+import CasePaths
 import ComposableArchitecture
 import UserNotifications
